pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "alpaca-py>=0.43.2",
    "numpy>=2.4.4",
    "pandas>=2.3.3",
    "plotly>=6.6.0",
    "pytz>=2026.1.post1",
    "streamlit>=1.55.0",
]
